==> 1-dc_preprocess.Rmd <==
---
title: "Load and pre-process all datasets in Gerhard et al. metanalysis"
author: "Massimo Andreatta & Santiago Carmona"
#output: html_notebook
---

```{r}
renv::restore()
```

Libraries
```{r}
library(ggplot2)
library(Seurat)
library(tidyr)
library(SeuratObject)
library(Matrix)
library(patchwork)
library(remotes)
library(UCell)
library(scGate)
```

# Preprocessing and data conversion

NB: for conversion between Scanpy format to matrix files, see export_data.py script

Read in count data for individual studies, standardize gene symbols, and store into a Seurat object for further processing.

Direct link to count matrices: https://www.dropbox.com/s/agr2r48723h44k8/exported.zip?dl=0
Direct link to pre-processed data: https://www.dropbox.com/s/uz590gzx4s36tdi/Gerhard_metanalysis_seuratlist.rds?dl=0

```{r}
#Change to your working directories
pdir <- "~/Dropbox/CSI/Datasets/Gerhard_JEM"
ddir <- "~/Dropbox/CSI/Datasets/Gerhard_JEM/exported"

geneRef_file <- "aux/EnsemblGenes105_Hsa_GRCh38.p13.txt.gz"
geneRef <- data.table::fread(geneRef_file)

studies <- list.dirs(path=ddir, full.names = FALSE)
studies <- studies[nchar(studies)>0]

names <- gsub("_\\S+","",studies, perl=T)

seurat.list <- list()

for (i in seq_along(studies)) {
  s = studies[i]
  proj = names[i]
   
  data <- t(readMM(file = sprintf("%s/%s/matrix.mtx", ddir, s)))

  cell.names <- read.table(sprintf("%s/%s/barcodes.tsv", ddir, s), sep="\t", header=T)[,1]
  gene.names <- read.table(sprintf("%s/%s/genes.tsv", ddir, s), sep="\t", header=T)[,1]
  meta <- read.table(sprintf("%s/%s/metadata.csv",ddir,s), sep=",",header=T)[,-1]
  rownames(meta) <- cell.names
  rownames(data) <- gene.names
  colnames(data) <- cell.names 

  #make Seurat object
  seurat.list[[proj]] <- CreateSeuratObject(counts=data, project=proj, assay="RNA", meta.data = meta)
  seurat.list[[proj]]$Study <- proj
  
  #make unique names
  id = paste0("S",i)
  seurat.list[[proj]] <- RenameCells(seurat.list[[proj]], add.cell.id = id)
  
  #standardize gene names
  seurat.list[[proj]] <- STACAS:::standardizeGeneSymbols(seurat.list[[proj]],
                                                         EnsemblGeneTable=geneRef)
  
  #mito and ribo content
  patterns <- c("^RP[LS]","^MT-")

  pc.ribo <- PercentageFeatureSet(seurat.list[[proj]], pattern = patterns[1])
  pc.mito <- PercentageFeatureSet(seurat.list[[proj]], pattern = patterns[2])
  
  seurat.list[[proj]] <- AddMetaData(seurat.list[[proj]], metadata = pc.ribo, col.name = "percent.ribo")
  seurat.list[[proj]] <- AddMetaData(seurat.list[[proj]], metadata = pc.mito, col.name = "percent.mito")
  
}

##And save the list of object to disk
saveRDS(seurat.list, file=sprintf("%s/Gerhard_metanalysis_seuratlist.rds",pdir))
```

Data were normalized using total-count normalization (https://www.sciencedirect.com/science/article/pii/S0092867415005000#app3). log1p normalization from NormalizeData should reverse the dataset-specific normalization factor.

# Basic exploration of data
```{r}
plots <- list()

ndim=30
for (i in seq_along(seurat.list)) {
  s = names(seurat.list)[i]
  
  obj <- seurat.list[[s]]

  obj <- NormalizeData(obj)
  obj <- FindVariableFeatures(obj, selection.method = "vst", nfeatures = 2000, verbose = FALSE)
  obj <- ScaleData(obj, do.scale = TRUE, do.center = TRUE, verbose = F) 
  obj <- RunPCA(object = obj, features = obj@assays$RNA@var.features, npcs = ndim, verbose = F)
  
  obj <- RunUMAP(obj, reduction = "pca", dims = 1:ndim, seed.use=123, verbose = F)
  plots[[s]] <- DimPlot(obj, reduction = "umap", group.by = "state", label = T) + ggtitle(s) + NoLegend() +
             theme(axis.title=element_blank(),axis.text=element_blank(),axis.ticks=element_blank())
}
```

```{r fig.height=8, fig.width=16}
wrap_plots(plots, ncol=4)

ggsave("plots/Gerhard_sets_unsup_umap.pdf", width = 24, height=15)
```


Focus on DCs only
```{r}
lapply(seurat.list, function(x) {unique(x$state)})

# brownSpleen, martin, qian, qianbc, qiancrc, qianlung, qianovarian are focused on DCs only
# for the rest, subset on DC subsets according to author annotations

dc.seurat <- seurat.list

dc.seurat$villani <- subset(dc.seurat$villani, subset=state %in% c("vi-DC1",
                                                                   "vi-DC5",
                                                                   "vi-DC2",
                                                                   "vi-DC3",
                                                                   "vi-DC6",
                                                                   "vi-DC4"))
dc.seurat$zhang <- subset(dc.seurat$zhang, subset=state %in% c("zhH-DC-C2-FCER1A",
                                                               "zhH-DC-C4-LAMP3",
                                                               "zhH-DC-C3-CLEC9A",
                                                               "zhH-DC-C1-CD1C" ))

dc.seurat$zhangcrc <- subset(dc.seurat$zhangcrc, subset=state %in% c("zhC-hM03_cDC2-CD1C",
                                                                     "zhC-hM04_cDC1-BATF3",
                                                                     "zhC-hM02_pDC-LILRA4" ))

dc.seurat$zilionis <- subset(dc.seurat$zilionis, subset=state %in% c("zi-tMonoDC",
                                                                     "zi-tDC1",
                                                                     "zi-tDC2",
                                                                     "zi-tDC3",
                                                                     "zi-tpDC"))

saveRDS(dc.seurat, file=sprintf("%s/Gerhard_metanalysis_seuratlist.DC.rds",pdir))
```

Assign main subtypes to common dictionary of "reference states", as defined by Gerhard et al.
```{r}
alt.names <- list()
alt.names[["cDC1"]] <- c("brS-cDC1","brS-Mitotic cDC1","ma-cDC1","mi-DC1",
                         "C1_CLEC9A","qiB-C1_CLEC9A","qiC-C1_CLEC9A","qiL-C1_CLEC9A",
                         "qiO-C1_CLEC9A","vi-DC1", "zhH-DC-C3-CLEC9A","zi-tDC1")

#alt.names[["cDC2"]] <- c("brS-CLEC10A+ cDC2","brS-CLEC10A- cDC2","brS-Mitotic cDC2",
#                         "ma-cDC2","mi-DC2","C2_CLEC10A","qiB-C2_CLEC10A","qiB-C5_CD207",
#                         "qiC-C2_CLEC10A","qiC-C5_CD207","qiL-C2_CLEC10A","qiL-C5_CD207",
#                         "qiO-C2_CLEC10A","qiO-C5_CD207","vi-DC2","C5_CD207",
#                         "zhH-DC-C1-CD1C","zhC-hM03_cDC2-CD1C","zi-tDC2")
#alt.names[["MonoDC"]] <- c("vi-DC3","vi-DC4","zi-tMonoDC","mi-moDC")

alt.names[["cDC2/MoDC"]] <- c("brS-CLEC10A+ cDC2","brS-CLEC10A- cDC2","brS-Mitotic cDC2",
                         "ma-cDC2","mi-DC2","C2_CLEC10A","C5_CD207","qiB-C2_CLEC10A","qiB-C5_CD207",
                         "qiC-C2_CLEC10A","qiC-C5_CD207","qiL-C2_CLEC10A","qiL-C5_CD207",
                         "qiO-C2_CLEC10A","qiO-C5_CD207","vi-DC2",
                         "zhH-DC-C1-CD1C","zhC-hM03_cDC2-CD1C","zi-tDC2",
                         "vi-DC3","vi-DC4","zi-tMonoDC","mi-moDC","zhH-DC-C2-FCER1A")

alt.names[["DC3"]] <- c("ma-mregDC","C3_CCR7","qiB-C3_CCR7","qiC-C3_CCR7",
                        "qiO-C3_CCR7","qiL-C3_CCR7","zhH-DC-C4-LAMP3",
                         "mi-Activated DC","zhC-hM04_cDC1-BATF3","zi-tDC3")

alt.names[["pDC"]] <- c("mi-pDC","C4_LILRA4","qiB-C4_LILRA4","qiC-C4_LILRA4",
                        "qiL-C4_LILRA4","qiO-C4_LILRA4","vi-DC6","zhC-hM02_pDC-LILRA4","zi-tpDC")

alt.names[["AS-DC"]] <- c("vi-DC5","brS-AS DC")

dc.seurat <- lapply(dc.seurat, function(x) {
  x$common.name <- NA
  for (n in names(alt.names)) {
    x$common.name[x$state %in% alt.names[[n]]] <- n
  }
  x
})

```

See original DC annotations from the different studies
```{r}
plots <- list()
plots.2 <- list()
ndim=15
for (s in names(dc.seurat)) {
  
  dc.seurat[[s]] <- NormalizeData(dc.seurat[[s]])
  dc.seurat[[s]] <- FindVariableFeatures(dc.seurat[[s]], selection.method = "vst", nfeatures = 500, verbose = FALSE)
  dc.seurat[[s]] <- ScaleData(dc.seurat[[s]], verbose = F) |> RunPCA(npcs = ndim, verbose=F)
  
  dc.seurat[[s]] <- RunUMAP(dc.seurat[[s]], reduction = "pca", dims = 1:ndim, seed.use=123, verbose = F)
  plots[[s]] <- DimPlot(dc.seurat[[s]], reduction = "umap", group.by = "state", label = T) + ggtitle(s) + NoLegend() +
             theme(axis.title=element_blank(),axis.text=element_blank(),axis.ticks=element_blank())
  
  plots.2[[s]] <- DimPlot(dc.seurat[[s]], reduction = "umap", group.by = "common.name", label = T) + ggtitle(s) + NoLegend() +
             theme(axis.title=element_blank(),axis.text=element_blank(),axis.ticks=element_blank())
}
```

```{r fig.height=6, fig.width=12}
#plots <- plots[names(plots) != "brownSpleen"]

wrap_plots(plots.2, ncol=4)
wrap_plots(plots, ncol=4)

ggsave("plots/Gerhard_sets_unsup_DC.pdf", width = 16, height=12)
```

# scGate DC subtype models

Can we build simple signatures based on marker genes?
```{r}
sigs <- list()

sigs$cDC1 <- c("CLEC9A","XCR1","BATF3","CADM1")
sigs$cDC2 <- c("CD1C","CLEC10A","FCGR2A","FCER1A")
sigs$DC3 <- c("LAMP3","CCR7","FSCN1")
sigs$pDC <- c("GZMB","LILRA4","TCF4")
#sigs$MoDC <- c("S100A9","S100A8","FCN1","VCAN")

dc.seurat <- lapply(dc.seurat, function(x) {
  dc.seurat <- AddModuleScore_UCell(x, features=sigs, name=NULL)
  SmoothKNN(dc.seurat, signature.names = names(sigs), reduction = "pca")
})
```

```{r fig.height=3}
for (s in names(dc.seurat)) {
  p <- FeaturePlot(dc.seurat[[s]], features=paste0(names(sigs),"_kNN"))
  plot(p)
}
```

Assign labels based on scGate (simple models based on a few genes)
```{r}
ss <- list()
ss$cDC1 <- gating_model(name="cDC1", signature = sigs$cDC1)
ss$cDC2 <- gating_model(name="cDC2", signature = sigs$cDC2)
ss$DC3 <- gating_model(name="DC3", signature = sigs$DC3)
ss$pDC <- gating_model(name="pDC", signature = sigs$pDC)
#ss$MoDC <- gating_model(name="MoDC", signature = sigs$MoDC)  #may be difficult to find a specific signature for MonoDC

dc.seurat <- lapply(dc.seurat, function(x) {
  scGate(x, model=ss, reduction = "pca", ncores = 4)
})
```


```{r fig.height=12, fig.width=4}
library(RColorBrewer)
palette <- brewer.pal(n=length(ss)+1, name="Set3")
names(palette) <- c(names(ss), "Multi")

pll <- list()
i <- 0
for (s in names(dc.seurat)) {
  pll[[i <- i+1]] <- DimPlot(dc.seurat[[s]], group.by = "scGate_multi", cols = palette) + theme(aspect.ratio = 1)
  pll[[i <- i+1]] <- DimPlot(dc.seurat[[s]], group.by = "state") + theme(aspect.ratio = 1) + ggtitle(s)
}
wrap_plots(pll, ncol = 2)

ggsave("plots/Gerhard_sets_scGate_multi.pdf", height=32, width=11)
```

Save DC datasets - with scGate annotations
```{r}
saveRDS(dc.seurat, file=sprintf("%s/Gerhard_metanalysis_seuratlist.DC.anno.rds",pdir))
```
