==> 2-dc_integrate.Rmd <==
---
title: "Integrate DC datasets"
author: "Massimo Andreatta & Santiago Carmona"
#output: html_notebook
---

```{r}
renv::restore()
```

Libraries
```{r}
#remotes::install_github('carmonalab/STACAS')
#remotes::install_github('carmonalab/scIntegrationMetrics')

library(ggplot2)
library(Seurat)
library(tidyr)
library(SeuratObject)
library(Matrix)
library(patchwork)
library(remotes)
library(UCell)
library(STACAS)
library(scIntegrationMetrics)
```


Read in data as Seurat objects (see 1-dc_preprocess.Rmd for data conversion)

Direct link to pre-processed data, with scGate annotations: https://www.dropbox.com/s/j50p3newch3hn8n/Gerhard_metanalysis_seuratlist.DC.anno.rds?dl=0
```{r}
pdir <- "~/Dropbox/CSI/Datasets/Gerhard_JEM"

seurat.list <- readRDS(sprintf("%s/Gerhard_metanalysis_seuratlist.DC.anno.rds",pdir))

#Exclude Brown dataset - we will use it as a validation set
seurat.list <- seurat.list[names(seurat.list) != "brownSpleen"]
```

See extent of batch effects
```{r fig.height=3}
merged <- Reduce(merge, seurat.list)

merged <- FindVariableFeatures(merged, nfeatures = 1000)
merged <- merged %>% ScaleData() %>% RunPCA(npcs=30) %>% RunUMAP(dims=1:30)

library(RColorBrewer)
palette <- brewer.pal(n=5, name="Set3")
names(palette) <- c("cDC1","cDC2","DC3","pDC","Multi")

#palette.2 <- brewer.pal(n=6, name="Set3")
#names(palette.2) <- c("cDC1","cDC2","DC3","pDC","MonoDC","AS-DC")
palette.2 <- brewer.pal(n=5, name="Set3")
names(palette.2) <- c("cDC1","cDC2/MoDC","DC3","pDC","AS-DC")

p1 <- DimPlot(merged, reduction="umap", group.by="Study") + theme(aspect.ratio = 1)
p2 <- DimPlot(merged, reduction="umap", group.by="scGate_multi", cols=palette) + theme(aspect.ratio = 1)
p3 <- DimPlot(merged, reduction="umap", group.by="common.name", cols=palette.2) + theme(aspect.ratio = 1)

p1 | p3

ggsave("plots/Gerhard_sets_preintegration.pdf", height = 5, width=10)
saveRDS(merged, "aux/DC_unintegrated.rds")
```


Find anchors
```{r}
nfeatures=2000
ndim=40
set.seed(1234)
stacas_anchors <- FindAnchors.STACAS(seurat.list, 
                                     anchor.features = nfeatures,
                                     dims = 1:ndim)
```

```{r}
tree <- SampleTree.STACAS(stacas_anchors,
                          obj.names = names(seurat.list),
                          hclust.method = "ward.D")
```

Integrate data
```{r}
object_integrated <- IntegrateData.STACAS(stacas_anchors,
                                          sample.tree=tree,
                                              dims=1:ndim)
```

Dimred
```{r}
set.seed(123)
object_integrated <- object_integrated %>% ScaleData() %>% RunPCA(npcs=ndim) %>% RunUMAP(dims=1:ndim)
```

```{r fig.height=3}
library(RColorBrewer)

p1 <- DimPlot(object_integrated, reduction="umap", group.by="Study") + theme(aspect.ratio = 1)
p2 <- DimPlot(object_integrated, reduction="umap", group.by="scGate_multi", cols=palette) + theme(aspect.ratio = 1)
p3 <- DimPlot(object_integrated, reduction="umap", group.by="common.name", cols=palette.2) + theme(aspect.ratio = 1)
p1 | p3

ggsave("plots/Gerhard_sets_STACAS_unsup.pdf", height = 5, width=10)
```


Unsupervised clustering
```{r}
k.param <- 20
res <- 0.2
set.seed(123)

object_integrated <- FindNeighbors(object_integrated, k.param=k.param, dims=1:ndim)
object_integrated <- FindClusters(object_integrated, resolution = res, graph.name = "integrated_snn")

DimPlot(object_integrated, reduction = "umap", group.by="seurat_clusters") + theme(aspect.ratio = 1)
```

Find markers for unsupervised clusters
```{r}
Idents(object_integrated) <- "seurat_clusters"
DefaultAssay(object_integrated) <- "RNA"

set.seed(1234)
cluster.markers <- FindAllMarkers(object_integrated, only.pos = T, min.pct = 0.1, min.diff.pct=0.1, 
                                  logfc.threshold = 0.25, max.cells.per.ident = 500, test.use="wilcox",base=exp(1))

all <- cluster.markers |> dplyr::group_by(cluster) |> dplyr::top_n(n = 50, wt = abs(avg_logFC))

for (i in levels(object_integrated@active.ident)) {
    print(subset(all, cluster==i))
}  
DefaultAssay(object_integrated) <- "integrated"

```


Where are specific subtypes, e.g. the "DC5" by Villani et al.?
```{r fig.height=10, fig.align=16}
p <- list()
anns <- unique(object_integrated$state)

for (a in anns) {
p[[a]] <- DimPlot(object_integrated, reduction = "umap", cells.highlight = which(object_integrated$state == a)) + 
  theme(aspect.ratio = 1) + NoLegend() + ggtitle(a)
}

wrap_plots(p)
```

```{r fig.height=3}
DefaultAssay(object_integrated) <- "RNA"

FeaturePlot(object_integrated, features=c("CLEC10A","IL1R2","CD1C","CFP"))
FeaturePlot(object_integrated, features=c("C1QA","C1QB","C1QC","FCGR2A"))
FeaturePlot(object_integrated, features=c("CD1A","CD207","FCGBP","S100B"))
FeaturePlot(object_integrated, features=c("S100A8","S100A9","FCN1","CD14"))
FeaturePlot(object_integrated, features=c("S100B","ACY3","PRDM16","CD14"))
FeaturePlot(object_integrated, features=c("LTB","RUNX3","LTB","IL22RA2"))
FeaturePlot(object_integrated, features=c("PPP1R14A","DAB2","AXL","LILRA4"))
FeaturePlot(object_integrated, features=c("TRAC","CD2","CD3E","CD3D"))
FeaturePlot(object_integrated, features=c("TOP2A","STMN1","MKI67","CD3D"))

```

Try to name clusters
```{r}
object_integrated$annotation <- NA
object_integrated$annotation[object_integrated$seurat_clusters %in% c("0")] <- "cDC2_CLEC10A"
object_integrated$annotation[object_integrated$seurat_clusters %in% c("1","9")] <- "cDC1"
object_integrated$annotation[object_integrated$seurat_clusters %in% c("2")] <- "DC3"
object_integrated$annotation[object_integrated$seurat_clusters %in% c("3","7")] <- "cDC2_CD1A"

object_integrated$annotation[object_integrated$seurat_clusters %in% c("4")] <- "MonoDC"
object_integrated$annotation[object_integrated$seurat_clusters %in% c("5")] <- "pDC"
object_integrated$annotation[object_integrated$seurat_clusters %in% c("6")] <- "Tcell_doublets"
object_integrated$annotation[object_integrated$seurat_clusters %in% c("8")] <- "AS-DC"

object_integrated$annotation[object_integrated$seurat_clusters %in% c("10")] <- "PRDM16_DC"
DimPlot(object_integrated, group.by = "annotation")
```

Make it into reference map
```{r}
library(ProjecTILs)
DefaultAssay(object_integrated) <- 'integrated'
ref <- make.reference(object_integrated, recalculate.umap = TRUE,
                      annotation.column = "annotation", ndim = 30)

```


```{r}
states <- levels(ref$functional.cluster)

palette <- brewer.pal(n = length(states), name = "Paired")
names(palette) <- states
ref@misc$atlas.palette <- palette

DimPlot(ref, group.by = "functional.cluster", label = TRUE, cols = palette) + theme(aspect.ratio = 1)
ggsave("plots/DC_reference_wTcells.pdf", height=5, width=7)
```

```{r fig.height=3.5}
DefaultAssay(ref) <- "RNA"
FeaturePlot(ref, features=c("CD1A","CD207","ENTPD1","FCGBP"))
FeaturePlot(ref, features=c("ACY3","PRDM16","LAMP3","TRAC"), order=T)
FeaturePlot(ref, features=c("LAMP3","BIRC3","GZMB","CCR7"))
FeaturePlot(ref, features=c("FCER1A","CD1C","CLEC10A","FCGR2A"))
```

```{r fig.height=3}
library(RColorBrewer)
ss <- unique(ref$scGate_multi)
ccc <- brewer.pal(n=length(ss)+1, name="Set3")
names(ccc) <- ss

p1 <- DimPlot(ref, reduction="umap", group.by="Study") + theme(aspect.ratio = 1)
p2 <- DimPlot(ref, reduction="umap", group.by="scGate_multi", cols=ccc) + theme(aspect.ratio = 1)
p3 <- DimPlot(ref, reduction="umap", group.by="common.name", cols=palette.2) + theme(aspect.ratio = 1)

p1 | p3

ggsave("plots/DC_reference_bysample.pdf", height = 5, width=10)
```

# Save object - reference with T cell doublets
```{r}
saveRDS(ref, "aux/DC_reference_wTcells.rds")
```


# Optional code below (additional analyses on current reference)


Find markers for subtypes in the reference
```{r}
Idents(ref) <- "functional.cluster"
DefaultAssay(ref) <- "RNA"

set.seed(1234)
cluster.markers <- FindAllMarkers(ref, only.pos = T, min.pct = 0.1, min.diff.pct=0.1, 
                                  logfc.threshold = 0.25, max.cells.per.ident = 500, test.use="wilcox",base=exp(1))

all <- cluster.markers |> dplyr::group_by(cluster) |> dplyr::top_n(n = 50, wt = abs(avg_logFC))

for (i in levels(ref@active.ident)) {
    print(subset(all, cluster==i))
}  
DefaultAssay(ref) <- "integrated"

```

Make a heatmap of DEGs
```{r fig.height=4}
ref.copy <- ref
DefaultAssay(ref.copy) <- 'RNA'
ref.copy <- ScaleData(ref.copy)

genes.use <- c()
for (i in levels(ref.copy$functional.cluster)) {
    this <- as.vector(all[all$cluster==i, "gene"])[[1]]
    this <- this[1:min(8, length(this))]
    
    genes.use <- c(genes.use, this)
}

heat <- DoHeatmap(ref.copy, assay="RNA", features=genes.use, group.by = "functional.cluster", group.colors = palette)

heat + scale_fill_gradientn(colors = c("blue", "white", "red"))
ggsave("plots/DC_scHeatmap_topDEG.png", height=11, width=21)
```

Subset composition
```{r}
tab <- table(ref$Study, ref$functional.cluster)
tab.norm <- tab/rowSums(tab)

df <- reshape2::melt(tab.norm)
colnames(df) <- c("Study","Subtype","Frequency")

palette <- ref@misc$atlas.palette

ggplot(df, aes(x = Study, y = Frequency, 
            fill = Subtype)) + geom_bar(stat = "identity") + 
            theme_bw() + scale_fill_manual(values = palette) +
            theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))
ggsave("plots/DC_composition_by_dataset.pdf", height=4, width=5)
```

Make a subset vs. study heatmap of pseudobulk profiles
```{r}
make.heatmap <- function(data, assay="RNA", genes, scale="row",
                         method=c("ward.D", "average"), brewer.palette="RdBu",
                         cluster.col = "functional.cluster", flip=FALSE,
                         cluster_genes = FALSE, cluster_samples=FALSE, min.cells = 10,
                         palette = NULL) {
  
  require(pheatmap)
  set.seed(123)
  
  method = method[1]
  
  data$Study_Subset <- factor(paste(data@meta.data[,cluster.col],data$Study,sep="."))
  
  t <- table(data$Study_Subset)
  accept <- names(t)[t>min.cells]
  
  data <- subset(data, subset=Study_Subset %in% accept)
   
  #Calculate mean expression by cluster
  m <- c()
  for( g in unique(genes)){
    m[[g]] <- tapply(data@assays[[assay]][g,],data$Study_Subset, mean)
  }
  m <- as.data.frame(m)
  
  m <- m[accept,]
  
  m.study <- factor(unlist(lapply(strsplit(rownames(m),"\\.",perl = T),function(x) x[[2]])))
  m.subset <- factor(unlist(lapply(strsplit(rownames(m),"\\.",perl = T),function(x) x[[1]])))

  #gaps_col <- (which(!duplicated(m.subset))-1)[-1]
  
  breaksList = seq(-2, 2, by = 0.1)
  
  require(RColorBrewer)
  color = colorRampPalette(rev(brewer.pal(n = 7, name = brewer.palette)))(length(breaksList))
  
  if (is.null(palette)) {
    palette = brewer.pal(n=length(unique(m.subset)), name="Paired")
    names(palette) <- unique(m.subset)
  }
  
  annotation_col = data.frame(
                    Subtype = m.subset,
                    Study = m.study
                )
  rownames(annotation_col) = rownames(m)


  if (flip) { 
    h <- pheatmap::pheatmap(m, cluster_rows = cluster_samples,
                            cluster_cols = cluster_genes,scale = scale,
                            breaks = breaksList, color=color, 
                            annotation_row = annotation_col, 
                            show_rownames = F,
                            border_color = NA,
                            annotation_colors = list(Subtype=palette), 
                            fontsize_row=6,fontsize = 7, 
                            clustering_method=method)
  } else {
    h <- pheatmap::pheatmap(t(m),cluster_rows = cluster_genes,
                            cluster_cols = cluster_samples,scale = scale,
                            breaks = breaksList, color=color, 
                            annotation_col = annotation_col, 
                            show_colnames = F,
                            border_color = NA,
                            annotation_colors = list(Subtype=palette), 
                            fontsize_row=6,fontsize = 7, 
                            clustering_method=method)
  }
  return(h)
}
```

Plot heatmap of pseudo-bulk by subtype/study
```{r}
library(pheatmap)
h <- make.heatmap(ref.copy, assay = "RNA", genes = genes.use, min.cells=15,
            palette=palette, cluster_genes = FALSE, cluster_samples = FALSE)

ggsave("plots/DC_heatmap_pseudobulk_unclustered.pdf", plot=h, height=6, width=9)

h <- make.heatmap(ref.copy, assay = "RNA", genes = genes.use, min.cells=20,
            palette=palette, cluster_genes = TRUE, cluster_samples = TRUE)
ggsave("plots/DC_heatmap_pseudobulk_clustered.pdf", plot=h, height=7, width=10)
```

# A deeper dive into T cell doublets. Are there preferences for DC subtypes and CD4/CD8 T cells?
```{r fig.width=4}
ref.copy2 <- ref.copy
DefaultAssay(ref.copy2) <- "RNA"
Idents(ref.copy2) <- "scGate_multi"

FeaturePlot(ref.copy2, features=c("cDC1_kNN","cDC2__kNN","DC3_kNN","Tcell_kNN"))
DimPlot(ref.copy2, group.by = "scGate_multi")

hist(ref.copy2$Tcell_kNN, breaks=30)

ref.copy2$is.doublet <- "FALSE"
ref.copy2$is.doublet[ref.copy2$Tcell_kNN > 0.2] <- TRUE

DimPlot(ref.copy2, group.by = "is.doublet")
ref.copy2 <- subset(ref.copy2, subset=scGate_multi %in% c("cDC1","cDC2","DC3"))
doub <- subset(ref.copy2, subset=is.doublet==TRUE)
nodoub <- subset(ref.copy2, subset=is.doublet==FALSE)

FeaturePlot(ref.copy2, features = c("CD3E","TRAC","CD8A","CD8B","CD4","CD40LG"), order=T)

a <- VlnPlot(doub, features = c("CD3E","TRAC","CD8A","CD8B","CD4","CD40LG"),  ncol=6)
b <- VlnPlot(nodoub, features = c("CD3E","TRAC","CD8A","CD8B","CD4","CD40LG"), ncol=6)

a / b
```

Smooth expression values, to mitigate sparsity
```{r fig.width=4}
genes <- c("CD3E","TRAC","CD8A","CD8B","CD4","CD40LG")

expr <- Seurat::GetAssayData(ref.copy2, assay="RNA", slot="data")[genes,]
expr <- t(as.matrix(expr))
ref.copy2 <- AddMetaData(ref.copy2, metadata=as.data.frame(expr))

ref.copy2 <- SmoothKNN(ref.copy2, signature.names = genes, k=5)

feats <- paste0(genes, "_kNN")
FeaturePlot(ref.copy2, features = feats, max.cutoff = 'q99', ncol=3)

doub <- subset(ref.copy2, subset=is.doublet==TRUE)
nodoub <- subset(ref.copy2, subset=is.doublet==FALSE)

a <- VlnPlot(doub, features = feats,  ncol=6, pt.size = 0)
b <- VlnPlot(nodoub, features = feats, ncol=6, pt.size = 0)

a / b

```

Violin plot of marker genes
```{r fig.height=4, fig.width=3}
DefaultAssay(ref.copy) <- "RNA"

split <- SplitObject(ref.copy, split.by = "Study")

genes <- c("PPP1R14A","AXL","CLEC9A","XCR1","CADM1","CD1A","CD207","CLEC10A","CD1C",
                 "LAMP3","CCR7","FSCN1","S100A8","S100A9","VCAN","FCN1",
                 "GZMB","LILRA4","TCF4","PRDM16","ACY3","ITGAX","TRAC","CD3E","CD3D","CD3G")

VlnPlot(ref.copy, features = genes, stack = TRUE, cols = palette, flip = TRUE, fill.by = "ident")

ggsave("plots/DC_annotated_violins.pdf", height=8, width=6)

VlnPlot(split$maier, features = genes, stack = TRUE, cols = palette, flip = TRUE, fill.by = "ident")
VlnPlot(split$qian, features = genes, stack = TRUE, cols = palette, flip = TRUE, fill.by = "ident")
VlnPlot(split$zilionis, features = genes, stack = TRUE, cols = palette, flip = TRUE, fill.by = "ident")
VlnPlot(split$qianlung, features = genes, stack = TRUE, cols = palette, flip = TRUE, fill.by = "ident")
VlnPlot(split$qiancrc, features = genes, stack = TRUE, cols = palette, flip = TRUE, fill.by = "ident")
VlnPlot(split$martin, features = genes, stack = TRUE, cols = palette, flip = TRUE, fill.by = "ident")

```
