==> 3-dc_integrate_removeTcells.Rmd <==
---
title: "Integrate DC datasets - exclude T cell doublets"
author: "Massimo Andreatta & Santiago Carmona"
#output: html_notebook
---

```{r}
renv::restore()
```

Libraries
```{r}
#remotes::install_github('carmonalab/STACAS')
#remotes::install_github('carmonalab/ProjecTILs')

library(ggplot2)
library(Seurat)
library(tidyr)
library(SeuratObject)
library(Matrix)
library(patchwork)
library(remotes)
library(UCell)
library(STACAS)
```

Read in pre-integrated object (see 2-dc_integrate.Rmd)
```{r}
ref.init <- readRDS("aux/DC_reference_wTcells.rds")

DimPlot(ref.init, group.by = "functional.cluster", label = TRUE, cols=ref.init@misc$atlas.palette) + theme(aspect.ratio = 1)
```

Violin plot of marker genes
```{r fig.height=5, fig.width=3}
DefaultAssay(ref.init) <- "RNA"

genes <- c("AXL","PPP1R14A","CLEC9A","XCR1","CADM1","CD207","CD1A","CD1C","CLEC10A","FCGR2A",
                 "LAMP3","CCR7","FSCN1","S100A8","S100A9","VCAN","FCN1","GZMB","LILRA4","TCF4",
           "PRDM16","TRAC","CD3D","CD3E","HLA-DRA", "HLA-DQA1","HLA-DQB1")

VlnPlot(ref.init, features = genes, stack = TRUE, cols = ref.init@misc$atlas.palette, flip = TRUE, fill.by = "ident")

ggsave("plots/DC_annotated_violins_wDoublets.pdf", height=9, width=6)

DefaultAssay(ref.init) <- "integrated"
```

Remove T cell doublets and cycling cells
```{r}
library(SignatuR)

ref.init$pre.label <- as.character(ref.init$functional.cluster)

#Simply remove T cell cluster
#ref.init <- subset(ref.init, subset=functional.cluster != "Tcell_doublets")
#ref.init$pre.label[ref.init$pre.label == "PRDM16_DC"] <- NA

DefaultAssay(ref.init) <- "RNA"
sigs <- list(Tcell=c("CD3D","CD3E","CD3G","TRAC"),
             cycling1=GetSignature(SignatuR$Hs$Programs$cellCycle.G1S),
             cycling2=GetSignature(SignatuR$Hs$Programs$cellCycle.G2M),
             ig=c("IGLC2","IGHG3","IGKC","IGHD","IGHG1","IGHA1","IGHG4","IGLC3","IGHGP","IGHM","IGHG2"))

ref.init <- AddModuleScore_UCell(ref.init, features = sigs, ncores=4)

ref.init <- SmoothKNN(ref.init, signature.names = c("Tcell_UCell","cycling1_UCell","cycling2_UCell","ig_UCell"))

FeaturePlot(ref.init, features=c("Tcell_UCell_kNN","ig_UCell","cycling1_UCell_kNN","cycling2_UCell_kNN"))

hist(ref.init$Tcell_UCell_kNN, breaks = 30)
hist(ref.init$cycling1_UCell_kNN, breaks = 30)
hist(ref.init$cycling2_UCell_kNN, breaks = 30)
hist(ref.init$ig_UCell_kNN, breaks = 30)

thr = 0.1
ref.init$toremove <- FALSE
ref.init$toremove[ref.init$Tcell_UCell_kNN>thr |
                  ref.init$cycling1_UCell_kNN>thr |
                  ref.init$cycling2_UCell_kNN>thr] <- TRUE
DimPlot(ref.init, group.by = "toremove")
```

```{r}
ref.init <- subset(ref.init, subset=toremove==FALSE)
ref.init <- subset(ref.init, subset=functional.cluster=="PRDM16_DC", invert=TRUE)

ref.init$pre.label[ref.init$functional.cluster=="Tcell_doublets"] <- NA
ref.init$pre.label[ref.init$functional.cluster%in% c("cDC2_CD1A","cDC2_CLEC10A")] <- "cDC2"

table(ref.init$pre.label, useNA = "ifany")
```

# Re-integrate with semi-supervised STACAS, using previous labels as a guide

Find anchors, guided by initial functional.cluster assignment
```{r}
DefaultAssay(ref.init) <- "RNA"
seurat.list <- SplitObject(ref.init, split.by = "Study")

nfeatures=2000
ndim=40
semi_supervised=TRUE  #whether to use labels to guide integration

if (semi_supervised) {
  ll <- "pre.label"
} else {
  ll <- NULL
}

set.seed(1234)
stacas_anchors <- FindAnchors.STACAS(seurat.list, 
                                     anchor.features = nfeatures,
                                     dims = 1:ndim,
                                     cell.labels = ll) 

tree <- SampleTree.STACAS(stacas_anchors,
                          obj.names = names(seurat.list),
                          hclust.method = "ward.D")
```

Integrate data
```{r}
object_integrated <- IntegrateData.STACAS(stacas_anchors,
                                          sample.tree=tree,
                                              dims=1:ndim)
```

Dimred
```{r}
set.seed(123)
object_integrated <- object_integrated %>% ScaleData() %>% RunPCA(npcs=ndim) %>% RunUMAP(dims=1:ndim)
```


```{r fig.height=5}
library(RColorBrewer)
palette <- brewer.pal(n=5, name="Set3")
names(palette) <- c("cDC1","cDC2","DC3","pDC","Multi")

#palette <- brewer.pal(n=6, name="Set3")
#names(palette) <- c("cDC1","cDC2","DC3","pDC","MoDC","Multi")


p1 <- DimPlot(object_integrated, reduction="umap", group.by="Study") + theme(aspect.ratio = 1)
p2 <- DimPlot(object_integrated, reduction="umap", group.by="scGate_multi", cols=palette) + theme(aspect.ratio = 1)
p3 <- DimPlot(object_integrated, reduction="umap", group.by="pre.label") + theme(aspect.ratio = 1)
p4 <- DimPlot(object_integrated, reduction="umap", group.by="functional.cluster") + theme(aspect.ratio = 1)

(p1 | p2) / (p3 | p4)
```

Recalculate clusters
```{r}
k.param <- 20
res <- 0.2
set.seed(123)


object_integrated <- FindNeighbors(object_integrated, k.param=k.param, dims=1:ndim)
object_integrated <- FindClusters(object_integrated, resolution = res, graph.name = "integrated_snn")

DimPlot(object_integrated, reduction = "umap", group.by="seurat_clusters") + theme(aspect.ratio = 1)
```

```{r}
Idents(object_integrated) <- "seurat_clusters"
DefaultAssay(object_integrated) <- "RNA"

set.seed(1234)
cluster.markers <- FindAllMarkers(object_integrated, only.pos = T, min.pct = 0.1, min.diff.pct=0.1, 
                                  logfc.threshold = 0.25, max.cells.per.ident = 500, test.use="wilcox",base=exp(1))

all <- cluster.markers |> dplyr::group_by(cluster) |> dplyr::top_n(n = 50, wt = abs(avg_logFC))

for (i in levels(object_integrated@active.ident)) {
    print(subset(all, cluster==i))
}  
DefaultAssay(object_integrated) <- "integrated"

```

Remove contaminants - IGs, erythrocytes, ...
```{r}
ref.init <- subset(object_integrated, subset=seurat_clusters %in% c(8,9,10), invert=T)
table(ref.init$pre.label)
```

```{r}
#Find anchors, guided by initial functional.cluster assignment
DefaultAssay(ref.init) <- "RNA"
seurat.list <- SplitObject(ref.init, split.by = "Study")

nfeatures=2000
ndim=30
semi_supervised=TRUE  #whether to use labels to guide integration

if (semi_supervised) {
  ll <- "pre.label"
} else {
  ll <- NULL
}

set.seed(1234)
stacas_anchors <- FindAnchors.STACAS(seurat.list, 
                                     anchor.features = nfeatures,
                                     dims = 1:ndim,
                                     cell.labels = ll) 

tree <- SampleTree.STACAS(stacas_anchors,
                          obj.names = names(seurat.list),
                          hclust.method = "ward.D")

object_integrated <- IntegrateData.STACAS(stacas_anchors,
                                          sample.tree=tree,
                                              dims=1:ndim)
```

```{r fig.height=5}
library(RColorBrewer)
palette <- brewer.pal(n=5, name="Set3")
names(palette) <- c("cDC1","cDC2","DC3","pDC","Multi")

set.seed(123)
object_integrated <- object_integrated %>% ScaleData() %>% RunPCA(npcs=ndim) %>% RunUMAP(dims=1:ndim)

p1 <- DimPlot(object_integrated, reduction="umap", group.by="Study") + theme(aspect.ratio = 1)
p2 <- DimPlot(object_integrated, reduction="umap", group.by="scGate_multi", cols=palette) + theme(aspect.ratio = 1)
p3 <- DimPlot(object_integrated, reduction="umap", group.by="pre.label") + theme(aspect.ratio = 1)
p4 <- DimPlot(object_integrated, reduction="umap", group.by="functional.cluster") + theme(aspect.ratio = 1)

(p1 | p2) / (p3 | p4)
```


Make it into reference map
```{r}
library(ProjecTILs)
DefaultAssay(object_integrated) <- 'integrated'
ref <- make.reference(object_integrated, recalculate.umap = TRUE, metric = "euclidean",
                      annotation.column = "functional.cluster", ndim = 40, seed=1234)

```



```{r fig.height=3}
states <- levels(ref$functional.cluster)
palette <- brewer.pal(n = length(states), name = "Paired")
names(palette) <- states
ref@misc$atlas.palette <- palette

DimPlot(ref, group.by = "functional.cluster", label = TRUE, cols = palette) + theme(aspect.ratio = 1)
ggsave("plots/DC_reference.pdf", height=5, width=7)
```

Re-cluster
```{r}
k.param <- 20
res <- 0.15
set.seed(123)


ref <- FindNeighbors(ref, k.param=k.param, dims=1:ndim)
ref <- FindClusters(ref, resolution = res, graph.name = "integrated_snn")
DimPlot(ref, reduction = "umap", group.by="seurat_clusters") + theme(aspect.ratio = 1)

```

```{r}
Idents(ref) <- "seurat_clusters"
DefaultAssay(ref) <- "RNA"

set.seed(1234)
cluster.markers <- FindAllMarkers(ref, only.pos = T, min.pct = 0.1, min.diff.pct=0.1, 
                                  logfc.threshold = 0.25, max.cells.per.ident = 500, test.use="wilcox",base=exp(1))

all <- cluster.markers |> dplyr::group_by(cluster) |> dplyr::top_n(n = 50, wt = abs(avg_logFC))

for (i in levels(ref@active.ident)) {
    print(subset(all, cluster==i))
}  
DefaultAssay(ref) <- "integrated"

```

```{r}
ref$functional.cluster <- NA
ref$functional.cluster[ref$seurat_clusters %in% c("0")] <- "cDC2_CLEC10A"
ref$functional.cluster[ref$seurat_clusters %in% c("1")] <- "cDC1"
ref$functional.cluster[ref$seurat_clusters %in% c("2")] <- "DC3"
ref$functional.cluster[ref$seurat_clusters %in% c("5","6")] <- "cDC2_CD1A"
ref$functional.cluster[ref$seurat_clusters %in% c("3")] <- "pDC"
ref$functional.cluster[ref$seurat_clusters %in% c("4")] <- "MonoDC"
ref$functional.cluster[ref$seurat_clusters %in% c("7")] <- "AS-DC"

ref$functional.cluster <- factor(ref$functional.cluster, levels=c("cDC1","cDC2_CLEC10A",
                                                                  "cDC2_CD1A","DC3","pDC",
                                                                  "AS-DC","MonoDC"))

ref@misc$atlas.palette <- ref@misc$atlas.palette[levels(ref$functional.cluster)]
Idents(ref) <- "functional.cluster"

DimPlot(ref, group.by = "functional.cluster", cols = ref@misc$atlas.palette, label=TRUE)

#Make a copy for calculating metrics
saveRDS(ref, "aux/DC_reference_2nd_wmeta.rds")
```


Add scGate model for filtering
```{r}
library(scGate)
models <- scGate::get_scGateDB(branch = 'dev', version='v0.11')

ref@misc$scGate$human <- models$human$generic$panDC
ref@misc$scGate$mouse <- models$mouse$generic$panDC
```

Remove unnecessary metadata
```{r}
ref@meta.data <- ref@meta.data[,c("orig.ident","Study","nCount_RNA","nFeature_RNA",
                               "barcode","numi","state",
                               "percent.ribo","percent.mito",
                               "cycling1_UCell","cycling2_UCell","functional.cluster")]

ref@misc$projecTILs <- "DC_human"
```



Save reference object
```{r}
pdir <- "~/Dropbox/CSI/Datasets/Gerhard_JEM"

saveRDS(ref, "aux/DC_human_ref_v1.rds")
saveRDS(ref, paste0(pdir,"/DC_human_ref_v1.rds"))
```


Make a down-sampled version for SPICA
```{r}
sdir <- "~/Dropbox/CSI/SPICA_atlases"

Idents(ref) <- "functional.cluster"
spica.ref <- subset(ref, downsample=2000)  #max 2000 cells per subtype
spica.ref <- DietSeurat(spica.ref, counts=FALSE, data=TRUE, dimreducs = c("pca","umap"))

DimPlot(spica.ref, cols = spica.ref@misc$atlas.palette)
table(spica.ref$functional.cluster)

saveRDS(spica.ref, paste0(sdir,"/DC_human_ref_v1_light.rds"))
```
