==> 6-dc_projection.Rmd <==
---
title: "Test reference-based analysis using the DC map"
author: "Massimo Andreatta & Santiago Carmona"
#output: html_notebook
---

```{r}
renv::restore()
```

Libraries
```{r}
library(ggplot2)
library(Seurat)
library(tidyr)
library(SeuratObject)
library(Matrix)
library(patchwork)
library(remotes)
library(UCell)
library(ProjecTILs)
```

Read in annotated reference map
```{r}
pdir <- "~/Dropbox/CSI/Datasets/Gerhard_JEM"
ref <- readRDS(paste0(pdir,"/DC_human_ref_v1.rds"))
```


```{r}
pdir <- "~/Dropbox/CSI/Datasets/Gerhard_JEM"

dataset.list <- readRDS(sprintf("%s/Gerhard_metanalysis_seuratlist.DC.anno.rds",pdir))

brown <- dataset.list$brownSpleen

DimPlot(brown, group.by = "state")
DimPlot(brown, group.by = "scGate_multi")
```

```{r}
brown.projected <- Run.ProjecTILs(brown, ref=ref, filter.cells = FALSE)
```



```{r fig.height=3.5, fig.width=10}
celltypes <- unique(brown.projected$state)
pll <- list()

for (c in celltypes) {
  sub <- subset(brown.projected, subset=state==c)
  pll[[c]] <- plot.projection(ref, query=sub, linesize = 0.5, pointsize = 0.5) + ggtitle(c)
}

wrap_plots(pll, ncol=4)
ggsave("plots/Brown_projected_bysubtype.noTcells.png", height=9, width=22)
```


Composition
```{r}
plot.statepred.composition(ref=ref, query=brown.projected)
ggsave("plots/Brown_projected_composition.noTcells.png", height=3.5, width=5)
```


Profile agreement
```{r fig.height=4}
genes4radar <- c("CLEC9A","XCR1","CADM1","CLEC10A","CD1C","CD1A","CD207",
                 "LAMP3","CCR7","FSCN1","S100A8","S100A9","VCAN","FCN1",
                 "GZMB","LILRA4","TCF4","PPP1R14A","AXL")

plot.states.radar(ref, query=brown.projected, genes4radar = genes4radar, min.cells = 10)
ggsave("plots/Brown_projected_radar.noTcells.png", height=10, width=14)
```


#How about mouse data?
```{r}
library(scRNAseq)

sce <- ZilionisLungData('mouse')

counts <- Matrix::Matrix(counts(sce),sparse = TRUE)
meta <- colData(sce)

seurat <- CreateSeuratObject(counts)
seurat <- AddMetaData(seurat, metadata = as.data.frame(meta))
```


```{r fig.width=6}
seurat <- seurat |> NormalizeData() |> FindVariableFeatures(nfeatures=1000) |>
  ScaleData() |>  RunPCA(npcs=20) |> RunUMAP(dims=1:20)

a <- DimPlot(seurat, group.by="Major.cell.type") + theme(aspect.ratio = 1)
b <- DimPlot(seurat, group.by="Minor.subset") + theme(aspect.ratio = 1)
a | b
```

Classifier in original space
```{r fig.width=6}
zilionis <- ProjecTILs.classifier(seurat, ref=ref, filter.cells = TRUE, ncores=8)

a <- DimPlot(zilionis , group.by="Major.cell.type") + theme(aspect.ratio = 1)
b <-  DimPlot(zilionis , group.by="functional.cluster") + theme(aspect.ratio = 1)
a | b
```

Projection in reference space
```{r}
zilionis.proj <- Run.ProjecTILs(seurat, ref=ref, filter.cells = TRUE, ncores=8)
plot.projection(ref=ref, query=zilionis.proj)
```

```{r fig.height=3.5, fig.width=10}
celltypes <- unique(seurat$Minor.subset)
celltypes <- grep('DC', celltypes, value=T)
pll <- list()

for (c in celltypes) {
  sub <- subset(zilionis.proj, subset=Minor.subset==c)
  pll[[c]] <- plot.projection(ref, query=sub, linesize = 0.5, pointsize = 0.5) + ggtitle(c)
}

wrap_plots(pll, ncol=4)
```

Profile agreement
```{r fig.height=4}
genes4radar <- c("CLEC9A","XCR1","CADM1","CLEC10A","CD1C","CD1A",
                 "LAMP3","CCR7","FSCN1","S100A8","S100A9","VCAN","FCN1",
                 "GZMB","TSPAN13","TCF4","PPP1R14A","AXL")

plot.states.radar(ref, query=zilionis.proj, genes4radar = genes4radar, min.cells =5)
```

Why are pDCs misclassified? poor human-mouse correspondence?
```{r fig.height=5}
DefaultAssay(zilionis.proj) <- "RNA"

t <- table(zilionis.proj$Minor.subset)
subs <- names(t[t>10])

tmp <- subset(zilionis.proj, subset=Minor.subset %in% subs)
Idents(tmp) <- "Minor.subset"

genes <- c("CLEC9A","XCR1","CADM1","CLEC10A","FCER1A","CD207",
                 "LAMP3","CCR7","FSCN1","GZMB","TCF4","PPP1R14A","AXL",
           "S100A8","S100A9","VCAN","FCN1","ITGAX")
genes <- genes[genes %in% rownames(tmp)]

VlnPlot(tmp, features = genes, fill.by = "ident")
```
Project the 'Tcell doublets' identified in first round of integration - can they be classified to their respective subtype?
```{r}
all.data <- readRDS("aux/DC_reference_wTcells.rds")
table(all.data$functional.cluster)
all.data$prev.label <- all.data$functional.cluster

dataset.list <- SplitObject(all.data, "Study")
```

```{r}
all.projected <- Run.ProjecTILs(dataset.list, ref=ref, filter.cells = FALSE,
                                ncores=8, fast.mode=TRUE)
```

```{r}
all.merged <- Reduce(f=merge.Seurat.embeddings, x=all.projected)
doublets <- subset(all.merged, subset=prev.label=="Tcell_doublets")

table(doublets$Study, doublets$functional.cluster)
```

```{r fig.height=4}
genes4radar <- c("CLEC9A","XCR1","CADM1","CLEC10A","CD1C","CD1A",
                 "LAMP3","CCR7","FSCN1","S100A8","S100A9","VCAN","FCN1",
                 "GZMB","TSPAN13","TCF4","PPP1R14A","AXL")

plot.states.radar(ref, query=doublets, genes4radar = genes4radar, min.cells =5)

plot.states.radar(ref, query=doublets, min.cells =5,
                  genes4radar = c(genes4radar,c("TRAC","CD8A","CD4") ))

```

Any preference for CD8 or CD4 in the doublets associated to different subtypes?
```{r fig.height=3}
t <- table(doublets$functional.cluster)
subs <- names(t)[t>10]
sb <- subset(doublets, subset=functional.cluster %in% subs)

DefaultAssay(sb) <- 'RNA'
VlnPlot(sb, features=c("CD4","CD8A","CD8B","CLEC9A","CLEC10A","VCAN"),
        group.by = "functional.cluster", pt.size = 0, cols = ref@misc$atlas.palette)
```

Compare to baseline

```{r fig.height=3}
nondoublets <- subset(all.merged, subset=prev.label=="Tcell_doublets", invert=T)

DefaultAssay(nondoublets) <- 'RNA'
VlnPlot(nondoublets, features=c("CD4","CD8A","CD8B","CLEC9A","CLEC10A","VCAN"),
        group.by = "functional.cluster", pt.size = 0, cols = ref@misc$atlas.palette)

a <- VlnPlot(sb, features=c("CD4","CD8A","CD8B"),
             group.by = "functional.cluster", pt.size = 0, cols = ref@misc$atlas.palette)
b <- VlnPlot(nondoublets, features=c("CD4","CD8A","CD8B"),
             group.by = "functional.cluster", pt.size = 0, cols = ref@misc$atlas.palette)
a / b
```
